[i3] Removed i3status replaced by polybar. Added exit on fail for lock

+++ files/i3/lock.sh
@@ -2,7 +2,15 @@
 
 tmpbg='/tmp/screen.png'
 
+# Needs scrot
 scrot "$tmpbg"
+
+[[ $? -neq 0 ]] && exit 1
+
+# Needs imagemagick package
 convert "$tmpbg" -spread 20 -implode 1 "$tmpbg"
+
+[[ $? -neq 0 ]] && exit 2
+
 i3lock -e -i "$tmpbg"
 rm "$tmpbg"
